// src/project/dto/create-project.dto.ts
export class CreateProjectDto {
    name: string;
    treasuryAddress: string;
    map: string;
    constructionTime: string;
    completionTime: string;
    managementForm: string;
    typeSource: string;
    desc: string;
    totalInvestment: string;
    place: [string];
    approvedInvestment: string;
    initInvestment: string;
    approvedInvestmentTime: any;
    initInvestmentTime: any;
    childProjects: any;
    investor: string;
    hasChildProject: boolean
}